+++ packages/manager/apps/dedicated/client/app/components/dedicated-cloud/dedicatedCloud.component.js
@@ -19,6 +19,7 @@ export default {
     vcdTrackingPrefix: '<',
     customerSurveyBannerAvailability: '<',
     goBackToList: '<',
+    customerVcfaasMigrationBannerAvailability: '<',
     isLogsDisabled: '<?',
   },
   controller,