+++ packages/manager/apps/pci-block-storage/src/utils/utils.spec.ts
@@ -1,0 +1,22 @@
+import { capitalizeFirstLetter } from '@/utils/index';
+
+describe('utils', () => {
+  describe('capitalizeFirstLetter', () => {
+    it.each([
+      [undefined, undefined],
+      [null, null],
+      ['', ''],
+      ['a', 'A'],
+      ['high-speed', 'High-speed'],
+      ['Toto', 'Toto'],
+      ['éléphant', 'Éléphant'],
+    ])(
+      'should capitalize the first letter, from %s to %s',
+      (input: string, expected: string) => {
+        const result = capitalizeFirstLetter(input);
+
+        expect(result).toEqual(expected);
+      },
+    );
+  });
+});