+++ packages/manager/apps/web-hosting/src/components/expirationDate/ExpirationDate.component.tsx
@@ -2,9 +2,9 @@ import { useParams } from 'react-router-dom';
 
 import { Trans, useTranslation } from 'react-i18next';
 
-import { OdsText } from '@ovhcloud/ods-components/react';
+import { Text } from '@ovhcloud/ods-react';
 
-import { useFormatDate } from '@ovh-ux/manager-react-components';
+import { useFormatDate } from '@ovh-ux/muk';
 
 import { useGetServiceInfos } from '@/data/hooks/webHostingDashboard/useWebHostingDashboard';
 import { SERVICE_INFOS_STATUS } from '@/data/types/product/ssl';
@@ -26,13 +26,13 @@ export default function ExpirationDate() {
       {!isExpired &&
         isInAutoRenew &&
         (data?.renew?.deleteAtExpiration ? (
-          <OdsText className="mb-7">
+          <Text className="mb-7">
             <Trans t={t} i18nKey="expiration_resiliation_date" values={{ expirationDate }}></Trans>
-          </OdsText>
+          </Text>
         ) : (
-          <OdsText className="mb-7">
+          <Text className="mb-7">
             <Trans t={t} i18nKey="expiration_renew_date" values={{ expirationDate }}></Trans>
-          </OdsText>
+          </Text>
         ))}
     </>
   );