+++ packages/manager/apps/dedicated/CHANGELOG.md
@@ -3,6 +3,36 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [20.70.0](https://github.com/ovh/manager/compare/@ovh-ux/manager-dedicated@20.69.0...@ovh-ux/manager-dedicated@20.70.0) (2025-12-18)
+
+
+### Features
+
+* **dedicated:** remove beta badge for logs to customer ([bb5bfcc](https://github.com/ovh/manager/commit/bb5bfcc1a2dc97d50f796eec85ff60496088274b)), closes [#MAOBS-165](https://github.com/ovh/manager/issues/MAOBS-165)
+
+
+
+
+
+# [20.69.0](https://github.com/ovh/manager/compare/@ovh-ux/manager-dedicated@20.68.10...@ovh-ux/manager-dedicated@20.69.0) (2025-12-18)
+
+
+### Features
+
+* **dedicated:** add migration vcfaas banner ([5615a55](https://github.com/ovh/manager/commit/5615a5505c1be188148306a160a37566f07c03a5)), closes [#MANAGER-19946](https://github.com/ovh/manager/issues/MANAGER-19946)
+
+
+
+
+
+## [20.68.10](https://github.com/ovh/manager/compare/@ovh-ux/manager-dedicated@20.68.9...@ovh-ux/manager-dedicated@20.68.10) (2025-12-18)
+
+**Note:** Version bump only for package @ovh-ux/manager-dedicated
+
+
+
+
+
 ## [20.68.9](https://github.com/ovh/manager/compare/@ovh-ux/manager-dedicated@20.68.8...@ovh-ux/manager-dedicated@20.68.9) (2025-12-16)
 
 **Note:** Version bump only for package @ovh-ux/manager-dedicated